[GUI] cs widget: minor usability improvements
Three simple fixes to the cold staking widget:

give immediate feedback (with red outline) when the owner address is not valid (same as we have with SendMultiRow objects, e.g. the staker address in this widget).

on send: try to unlock the wallet only when all validity checks are passing.

since the address shown on the confirmation dialog is the staker, change the label "Sending to" in "Delegating to".