[Cleanup] Remove obsolete multisig code
Remove the backend code which was added (in # 278) to support the management of multisig addresses via GUI. Since our current GUI no longer has this feature, this code is currently disconnected and, therefore, can be cleaned up.

If we ever decide to re-add a GUI tool for multisig addresses, we can re-implement it properly.