[Travis] Bump macOS CMake target image
A previous bug in the TravisCI macOS images required us to force a full
homebrew update rather than rely on pre-existing package versions.
TravisCI has now updated their macOS images to fix the bug, but only for
11.x+ versions.

Further, by doing a full homebrew update, the test environment updated
to the most recent Qt 5.15 version, which we do not yet fully support (
[GUI] Add missing QPainterPath include  fixes this),
and was causing the build to fail.

We can go with the 11.1 image and revert to update: false for now,
which gives us Qt 5.14.1.