[GUI] Fix dashboard txs list blinking issue
The batch processing mode of the QListView is the cause of the list blinking on every update.

Seems that internally, QT fires up a screen region update that it's not calculated properly when the batch mode is enabled. Forcing to update the whole view even when there is no visual change for the rows that are being shown.