[Wallet] Abandon tx in CommitTransaction if ATMP fails
It takes the opposite approach.
When ATMP fails in CommitTransaction, instead of just returning the txid (hoping that the failure is not permanent) try to directly abandon the transaction.
Save the result of these operations in a new struct CWallet::CommitResult, and use it to give a detailed feedback to the user (via log lines, JSON-RPC errors and GUI dialogs).

Based on top of:

 CValidationState logging