[GUI] Load persisted transaction filter during start
The transaction filter settings were persisted, but never loaded when the wallet starts.

Since it would be tricky to compute the index of an item of the filter-combobox from the filter value this change does it in the other direction, it persists the index of the item and computes the filter value from it which is trivial.