[GUI] Solving old, not loaded at startup, transactions notification issue
Problem: we are still signaling updates from really old transactions from the backend and, at the same time, the GUI is not loading more than 20k transactions into the transactionTableModel.

So, old transactions that are not being shown deliberately in the GUI are re appearing due the signaling update process. Increasing the cached model size again, data being shown in the GUI and erroneously popping up the new tx ballon notification.

This PR is solving it with the following changes:
Storing the first loaded transaction time and not showing signaled transactions that have time smaller than the first tx loaded time.
Only applied if the wallet is exceeding the maximum loaded transactions limit in the GUI (20,000 transactions).