[Model] Wallet interface refactor + code cleanup.
This is motivated by the different types of balances that we currently have (and the new ones that are coming).

Refactor: Back ported and adapted the WalletBalances wrapper class.
Cleanup: Removed the not used balance getters from walletModel + some not implemented methods that we had.
Cleanup: Removed the zerocoin balance related fields from the GUI, we are not showing it anymore and the GUI was still calculating + caching them.