Wallet: wtx cached balances test coverage + getAvailableCredit problem fix.
Added test coverage for the CWalletTx balances cache structure to verify the correctness of CREDIT/ISMINE_SPENDABLE, DEBIT/ISMINE_SPENDABLE, AVAILABLE_CREDIT/ISMINE_SPENDABLE, and a 'fake' block creation flow to verify CWallet::GetUnconfirmedBalance. (Leaving for a future PR, the cold staking balance types coverage)
Plus, popped up an issue in CWalletTx::GetAvailableCredit that is solved too.

Issue description:
As is now, CWalletTx::GetAvailableCredit is caching the amount using the AmountType::CREDIT when need to be using AmountType::AVAILABLE_CREDIT.
As them are different type of balances and are using the same cache position, only one of those is stored and the other one returns, invalidly, the value of the first one cached.

I intentionally have committed the unit test before the fix. So it can be tested easier with and without the fix. Placing the repo's head at 7f45afd, the wallet_tests.cpp will fail due this issue. Then can move to the next one (or directly to the last one) and it will successfully pass.