[Travis] Lower the build timeout for the functional tests job
in order to increase the time reserved for executing the tests.

Currently the 8th travis job fails, and saves a cache of the binaries, at the end of the build phase, if it took more than 14 minutes.
This leaves at least 36 minutes to run the functional test suite, which is not enough anymore.
Lower the build timeout to 9 minutes, so we have at least 41 minutes for the tests. We'll need to restart that job more often, but we'll see less timeouts.