[Core] Per-txout model for chainstate database
The chainstate database, and its cache, rely on a per-tx model (i.e. it is a map from txids to lists of unspent outputs).
This PR chages it to a per-txout model (i.e. a map from outpoints to single unspent outputs).

Pros:

Simpler code.
Avoiding the CPU overhead of deserializing and serializing the unused outputs.
More predictable memory usage.
More easily adaptable to various cache flushing strategies.

Cons:

Slightly larger on-disk representation, and sometimes larger in-memory representation (when there are multiple outputs for the same txid in the cache, which becomes optional).

Adapted from bitcoin # 10195 (adding a fCoinStake boolean memeber to the Coin class and considering BIP34 always in effect as per
[Core] Remove BIP30 check)

Based on top of:

[Core] Remove UTXO cache entries when the tx they were added for is removed
[P2P] Improve AlreadyHave
[Refactoring] Break circular dependency main ↔ txdb
[BUG] Properly copy fCoinStake memeber between CTxInUndo and CCoins
[Core] Use unordered_map for CCoinsMap and fix empty vectors in streams
[Core] Safer modify new coins
[Core] per-txout model preparation

The actual PR starts with "Introduce Coin, a single unspent output" (171bf0b)

⚠️ Note for testers: Do a backup of the chainstate before testing. With this PR, at startup, the database is automatically upgraded to the new format (so a reindex is not needed).