[Bug] wrong reserveKey when committing budget/proposal collaterals
The CReserveKey passed to GetBudgetFinalizationCollateralTX/GetBudgetSystemCollateralTX (for the change address) is not the same that is later passed to CommitTransaction.
So, if commit fails, for whatever reason, the original key is not returned to the keypool.

Also, the implementations of GetBudgetFinalizationCollateralTX and GetBudgetSystemCollateralTX are absolutely identical. The only difference is a single constant variable (the fee amount used).
Unify them in CreateBudgetFeeTX (passing a bool fFinalization to select the correct proposal/finalized-budget fee amount).