Move coinbase & coinstake to P2PKH
Moved the last remaining P2PK scripts to P2PKH (coinbase and the staking output to the current P2PKH standard).

The main argument used for the migration is that we are already using it as a standard everywhere in the sources (same as upstream) and has an increased, theorical, security due the pub key not being revealed initially (scriptsig reveals it).