[GUI] Start masternodes polling timer only when is needed
The masternodes polling was starting when the clientModel object is generated, locking cs_main every certain amount of time without care on whether the information is being presented or not.
This PR changes it, the polling timer will only be started when the information is needed on the screen and stopped when the user leaves such screen.