[BUG] Properly copy fCoinStake memeber between CTxInUndo and CCoins
The fCoinStake variable of the undo metadata, which is constructed in UpdateCoins, is never set.
Similarly, the coins cache entry is not restoring the fCoinStake flag from the undo data in ApplyTxInUndo.

Now, since coins->IsCoinStake() is checked within places like CheckTxInputs/AcceptableInputs/etc..., this could potentially lead to consensus issues in case of reorganizations.