[CMake] Define MAC_OSX for cmake builds on macOS
A few places in our code rely on this precompiler definition, most
notably in the datadir and zk-params dir default values. Autotools sets
this automatically, but it is not passed into encocoin-config.h as it
isn't necessary for autotools builds.

Instead of modify the autotools configure script to pass this define
into the config file, simply define it in CMakeLists.txt

and some fixes