[Core] Per-txout model for chainstate database
The chainstate database, and its cache, rely on a per-tx model (i.e. it is a map from txids to lists of unspent outputs).
This PR chages it to a per-txout model (i.e. a map from outpoints to single unspent outputs).

Pros:

    Simpler code.
    Avoiding the CPU overhead of deserializing and serializing the unused outputs.
    More predictable memory usage.
    More easily adaptable to various cache flushing strategies.

Cons:

    Slightly larger on-disk representation, and sometimes larger in-memory representation (when there are multiple outputs for the same txid in the cache, which becomes optional).

Adapted from bitcoin # 10195 (adding a fCoinStake boolean memeber to the Coin class and considering BIP34 always in effect as per # 1775)

Based on top of:

# 1799
# 1800
# 1797
# 1796
# 1795
# 1793

    # 1773

The actual PR starts with "Introduce Coin, a single unspent output"

⚠️ Note for testers: Do a backup of the chainstate before testing. With this PR, at startup, the database is automatically upgraded to the new format (so a reindex is not needed).