[Miner] Unifying the disperse coinbase tx flow + further clean up.
Another step forward improving the miner CreateNewBlock function.
This time, have unified the coinbase tx creation flow that was previously dispersed over the function and having some redundant initializations/sets.
Plus, cleaned up some unused and/or unneeded fields and function arguments.

This is a preparation for # 1815 work. It will not be possible to modify any of the elements of a transaction once them are appended to a block, thus why the transaction needs (more than ever) to be crafted in one single place and not all over the sources.