[Miner] PoS process, calculate stakeable utxo only once.
Work on top of # 1816 and part of # 1817 coinstake transaction creation speedup goal. Starting on 018 14d5

Improving the following situation: we are calculating the stakeable utxo twice for the same PoS block creation process (looping over the entire wallet's transactions map twice), one before calling CreateNewBlock in CheckForCoins and the second one inside the CreateNewBlock method when we call CWallet::CreateCoinStake.
This PR fixes it adding an unique available coins vector calculated before the block creation, only once per try, in the CheckForCoins function and feeding CreateNewBlock with it.