[Cleanup] Removing unused and unneeded functions and members
Going further with the cleanup, have removed several unused and not needed functions and files from the GUI and the wallet. The LongDescriptionRole (which was calling to the TransactionDesc class) is not needed anymore, just a remanent of the previous GUI. Then have gone down the rabbit hole and cleaned the methods and fields that were connected to the removed function and currently are not needed anymore.