[Refactor] Emplace elements in maps and vectors
With C++11, variadic templates and perfect forwarding offer a new way of adding elements into a container by means of emplacing (creating in place).

    https://en.cppreference.com/w/cpp/container/map/emplace
    https://en.cppreference.com/w/cpp/container/vector/emplace_back

This is particularly useful.
With the traditional insert()/push_back() we are likely to make unnecessary copies (or move); for example, with maps, first the std::pair<const K, V> is created, and then it's copied to the container.
With emplace(), instead, references to the key/value objects are forwarded directly to the constructor of the value_type inside the data structure without making any copy of std::pair<const K,V>.

This PR replaces most instances of:

    .insert(std::make_pair(K,V)) with .emplace(K, V)
    .push_back(Class(Args...)) with .emplace_back(Args...)

in the whole code base.
It touches many files (therefore is surely conflicting with many open pull requests), but changes are trivial.