[RPC] Update univalue and unify JSON obj pushes
The pervious usage of push_back(Pair()) was needed for compatibility with the old json spirit library, which we transitioned away from quite a while ago in favor of Univalue.

Univalue provides a more convenient method for pushing key/value pairs into a JSON object or array in the form of pushKV(), which is easier to read and leads to shorter code segments.

This PR first updates our univalue subtree with upstream, which has removed the now-unneeded declaration methods for push_back(Pair()), as well as patch a bug with pushKV() being supplied with a boolean instead of an integer argument.