[CMake] Add test_pivx unit test target
This adds support for compiling the unit test binary test_encocoin
directly with CMake. This also allows for CLion to perform code analysis
and completion on the test source files.

Requires the hexdump utility to be installed.

Note: actually running the unit tests from within CLion (Boost.Test targets) is not optimized, and the tests take a considerable amount of time compared to running the test_encocoin binary.