[Refactoring] Budget, round 2: review locks
Keep going with the masternode-budget classes refactoring, mainly addressing the thread synchronization aspect.

There is currently a potential deadlock due to wrong lock order between cs_main and cs_budgets, which will be solved in a follow-up PR as it needs a more invasive refactoring (decoupling the checks in UpdateValid/IsCollateralTransactionValid, doing those that don't depend on the chainstate just once, and not requiring cs_main held for the others).