BugFix: g_IsSaplingActive flag is not initialized at startup.
Small bug introduced in # 1814, the flag is set on every new tip but not initialized at startup.
Discovered rebasing # 1798 on top of master and running any sapling functional test. As g_IsSaplingActive is false at startup, the transaction hash is different, making the merkle tree hash different, failing in VerifyDB.