[RPC][BUG] Fix 'mnfinalbudget show' returning a single budget
Fixes a bug in the RPC mnfinalbudget show.

All finalized budgets have the same name ("main").
This is used as key for the object returned by the RPC, therefore only one single finalized budget is returned (despite there being many more in the map).

Fix it by adding the hash to the object key (removing it from the object value).

Side note: why finalized budgets do even need a strBudgetName (which is the same for every one) is a mystery to me... but, since it's part of the broadcast, we can't just outright remove it, it needs to be guarded with a network upgrade (future work).