[BUG][GUI] Fix wallet crashing on faq-buttons press
Bug reported today on discord.
Wallet (GUI 4.3.0) crashes when the FAQ buttons from the masternode widget are pressed.

Cause:
The FAQs sections have been recently reduced to only 6 (old zerocoin sections have been removed).
But the buttons in the masternode widget still reference the old indexes (9 and 10), which are now out of bounds.
Further, the safety check inside setSection, still has the old (10) max value hardcoded, so does not prevent this.

Fix:
This clearly shows the difficulty in maintaining code with fixed magic numbers (which depend on a list of objects that may be modified in future releases).
So, let's introduce a proper enumeration inside SettingsFaqWidget, and use that to manage the sections, instead of just integers.