[Bug] Fix misaligned height in calls to GetBlockValue
This is a known old bug, which was never truly fixed, due to it being harmless (and discovered after the last block-value reduction).

The function GetBlockValue(int nHeight) returns the block-reward value for a block at a given height (based on a "halving" schedule hardcoded within the function).
The issue is that it was being called passing the height of the previous block.
At some point, one particular call (in CreateCoinStake) was changed to use the (correct) current block height, while leaving the wrong one in all other places (including masternode payments/budgets and validation code).