[Bug][RPC] Guard RPC server during warm-up
Problem:
Most RPC commands crash the wallet during the initialization phase.

Solution:
An attempt to fix this specifically for getinfo was made in # 543, checking the existence of chainActive.Tip().
A better solution, instead of checking all places that try to access the block index, (or the chainstate, or the second layer maps, etc...), is to throw a JSON-RPC exception with any call, if the server is still in warm-up phase (returning the exact init message), directly from CRPCTable::execute.