[RPC] Cache money supply on memory. Introduce getsupplyinfo
With # 1878 the RPC call getinfo becomes awfully slow (it triggers a flush of the chainstate to disk and recalculates the amount).

In this PR we introduce a in-memory global object (instance of a dumb storage class) to cache the money supply, as well as the height of the chain when it was last updated.
The cache is updated in 3 cases:

    at startup
    after periodic full flushes to disk (blocks and chainstate), except during shutdown or IBD
    on-demand via RPC (see below)

A new RPC getsupplyinfo is introduced. It returns the contents of the cached object, after an optional flush/recalculation (if the forceupdate argument is set to true).

The money supply is still part of getinfo output, for backward compatibility, but the fetching is redirected to the new getsupplyinfo (with default value, false, for the forceupdate argument, thus without flush/recalculation).