[Refactoring] Budget, round 3: split UpdateValid checks for proposals / finalized budgets
Based on top of:

# 1843

# 1844

Current PR starts with [Refactor] Decouple CBudgetProposal::UpdateValid checks (6c6 a97e).

This focuses on extracting context-independent checks and doing them only once in AddProposal/AddFinalizedBudget.
As for the confirmation/expiration check (which is the only thing left for UpdateValid) , it is done without locking cs_main (while holding cs_budgets/cs_proposals) by caching the height of the block with the collateral transaction and relying on it (todo: it needs to be addressed the case of it being reorganized from the chain).

This new field, called nBlockFeeTx is added directly to the serialization of CBudgetProposal and CFinalizedBudget, so the old budget.dat will be invalidated and recreated new at startup.