[Validation] Introduce rolling cache for block hashes in masternode manager
Problem:
Validation of many objects on the second layer (masternode pings, winners/scores, etc...) requires access to chainActive to get the hashes of blocks at a certain height. This is done very often (especially for masternode pings) and means holding cs_main lock every time, making the process slow and possibly leading to locking order inconsistencies.

Solution:
Cache the last N block hashes in the masternode manager.
In order to implement this efficiently, we introduce a new data structure called CyclingVector, which is a fixed-size vector, exposing only a setter Set(i, value) which modifies the index i % N.
The cache is initialized at startup and updated when blocks get connected to / disconnected from the tip.

The size (CACHED_BLOCK_HASHES) is currently set a default value of 200 making it possible to validate all masternode pings (but not all masternode winners) without accessing chainActive. In the future we could add a startup flag to let the user customize this (higher cache size would, as usual, mean faster execution but higher memory usage).

An RPC getcachedblockhashes is introduced, in order to print the content of the cache. This can be used later in the functional test suite for better testing.

This PR also removes the accesses to chainActive to get the current chain-height (using a value cached in the manager).