[Refactoring] Budget, round 4: remove Broadcast classes
Continues the budget-refactoring thread.

Remove the following classes entirely

    CBudgetProposalBroadcast
    CFinalizedBudgetBroadcast

Replace them with a new dastream-based constructor and a GetBroadcast function (returning the serialized stream) inside CBudgetProposal / CFinalizedBudget classes.

This allows us to get rid also of mapSeenProposals and mapSeenFinalizedBudgets and all the relative ugly code duplication (just rely on mapProposals and mapFinalizedBudgets to check for existence, and to get the broadcast messages to send).