[Travis] Set stricter timeout for Dash 32bit job
The unit tests for this job take upwards of 35 minutes due to 32-bit
performance restrictions, so set a stricter build timeout.

A build timeout of 600 seconds (10 minutes) was chosen as it is roughly double what a (currently) fully ccache'd build requires, so some leeway is left over to avoid unnecessary job restarts when a PR's changeset has a minimal impact on the build time.

Additionally, revert the workaround "hack" to continuously output .'s to the console when running the unit tests as # 1944's reordering to run longer tests towards the beginning of the process has made it unnecessary. This also restores the easier to read output formatting.