[Refactoring] Budget, round 5: proposals/budgets FeeTX indexes
Starting with [Budget] Introduce mapFeeTxToProposal and mapFeeTxToBudget (0 eb 9e 45)

Since # 1845 we no longer check repeatedly the active chain, when updating a budget object, we just save the height of the block including the collateral tx (ref # 1845), when we first add the proposal/budget to the map.

Therefore we need to handle the case where said block is disconnected from the chain.

In order to do so efficiently, this PR adds two indexes (mapFeeTxToBudget/mapFeeTxToProposal), mapping collateral txids to the relative budget objects (while such objects are stored in the map).
A simple function CBudgetManager::RemoveByFeeTxId() can then be called from DisconnectBlock when undoing transactions, in order to remove the now-conflicted budget objects (without performing any expensive search).