[Core] Say "Goodbye" to SwiftTX
SwiftTX has been disabled via spork ages ago, and no work has been done on it since.
It will never be re-introduced in its current form.
Later on (after the introduction of deterministic masternodes and LLMQ), we should work on a DIP10-style design, implemented from scratch.

So... better not to have all this dead-code spread around, only providing another place for bugs to hide.
It's time to say "goodbye".