[Core] Sapling transaction version
Uniform the definition of isSapling transaction, which originally was meant only for txes with version equal to 2, and it was later changed to include also txes with version greater than 2 (but not updated everywhere).

Add a IsShieldedTx utility function defining "shielded tx" any transaction with isSapling version and with non-empty sapData.

Use GetTotalSize to get the serialized size of a transaction (instead of calling GetSerializeSize with CTransaction::CURRENT_VERSION).

Remove the possibility of having non-standard txes due to their version, after v5 enforcement.
Transactions with invalid version are now directly rejected in CheckTransaction.

Re-enables the IsStandardTx check on regtest.