[Wallet] Shielded transactions balance cache system.
Up until now, one of the several capabilities that # 1798 introduced was the wallet's shielded transactions balances calculation, a purely on-demand process (which is not so performant as txs balances are requested more than once during the entire software lifecycle).
This PR is going further, introducing a new wallet balance cache system for shielded transactions. Covered by a good amount of units tests that can be run pre and post the new cache system.

This PR is obviously built on top of 1798. Part of the shielded GUI work that will be coming soon.