[BUG] Fix transaction size limits for sapling version
The size limits for a sapling transaction are currently checked in 3 places:
1- CheckTransaction (validation.cpp)
2- CheckTransactionWithoutProofVerification (called by SaplingValidation::CheckTransaction)
3- SaplingValidation::ContextualCheckTransaction

While in (2) and (3) the size is correctly checked against MAX_TX_SIZE_AFTER_SAPLING, in (1) it is checked against the old MAX_ZEROCOIN_TX_SIZE.
So a shielded transaction with size greater than MAX_ZEROCOIN_TX_SIZE would be wrongly rejected in CheckTransaction.

Fix the check in (1), selecting the correct max size for shielded/non-shielded transactions.
Move the size-check before calling SaplingValidation::CheckTransaction (as it's faster).
Remove the checks in (2) and (3) that are just redundant.