[Cleanup] Stop using Boost where possible
This removes the dependency on Boost for a number of things that either have a direct c++11 equivalent, or could have a custom function dropped in with minimal effort. Also removes a number of unused boost header includes.

I didn't replace Boost's scoped_ptr with c++11's scoped_ptr as that is already being done in # 1943.

I also decided to not replace boost::assign::[map_]list_of here as its more invasive and better suited for it's own PR.