[RPC][Tests] Add sapling mempool test + getbestsaplinganchor
This adds a functional test to check the bahavior of sapling transactions with the mempool.
Specifically, it verifies the nullifiers/anchor management introduced in # 1958:

    If a transaction spends a sapling note, already spent by another in-mempool transaction, it's not accepted in the mempool.
    Same if it tries to double-spend a note already spent on chain.
    If, after disconnecting a block, the sapling merkle tree root changes, any transaction that spends a note referencing the old root must be evicted from the mempool.

In order to verify the last point, a simple RPC getbestsaplinganchor is introduced in the "blockchain" category.
It returns the pcoinsTip best anchor.

Note: without 1958, the test crashes the node, as two transactions spending the same note can both enter the mempool, but then the assertions in the memory pool consistency checks fail during block creation.