[Wallet][DB] Back ports up to the CWallet/CWalletDB/CDB refactor.
Upgrading the wallet, wallet db and db wrapper areas with several back ports and updates adaptations.

List of PRs:

    bitcoin # 5933
    bitcoin # 7537
    bitcoin # 8191
    bitcoin # 8564
    bitcoin # 8574
    bitcoin # 8629
    bitcoin # 9353

Note:
Low priority PR, can be reviewing after merge Sapling and tier two PRs.
I made it because we are pretty far from upstream in the db area and want to introduce some new functionality in the future that requires a more up-to-date db layer.