Coin control dialog utxo processing/update flow large cleanup and improvements
Base initial work in order to be able to generalize CoinControlDialog up to the point of introducing shielded notes selection (being developed in the Sapling GUI PR # 1963).

Found several redundancies and not so good code over the overall CoinControlDialog class. So.. have re-structured it in a large percentage, aiming to improve the performance and clean all of the inefficiencies.