[Consensus][Validation] Dynamic Fee for Shielded Transactions + Max Size.
This PR proposes a mitigation for the attack known as "Sapling Woodchipper" (CVE-2019-11636).
Some references:

    https:// nvd.nist. gov/vuln/detail/CVE-2019-11636
    zcash/zcash # 3983

Here we:

    Define a dynamic minimum fee for shielded txes, based on the value of minRelayTxFee and the transaction size.
    The minRelayTxFee is multiplied by a factor K (currently set to 1000, and in the future could be variable based on the number of transactions in the mempool).

With the default value for minrelaytxfee (0.0001 PIV/kB), a shielded tx costs 0.0064 PIV plus:

    0.0148 PIV per transparent input (P2PKH spend)
    0.0384 PIV per shielded spend
    0.0034 PIV per transparent output
    0.0948 PIV per shielded output

A simple shielded-to-shielded transaction (with shielded change) thus, for example, requires a minimum fee of 0.2344 PIV (~0.06 USD at the time of writing). A "mint" transaction (transparent to shielded, with transparent change) requires 0.1194 PIV.
Transactions with a fee higher than 100 times the minimum fee are not accepted in the mempool.

    Update the RPC shieldedsendmany/rawshieldedsendmany to automatically compute the minimum possible fee as default, when the fee parameter is not set.

    Lower the maximum allowed size for a shielded transaction to 400 kB.
    It would require ~40 PIVs in fees, and in order to fill a block, 5 of such transactions would be needed.

The total cost of the attack is thus raised (from 0.0001 PIV/min) to 200 PIV/min (~3600 USD/h).

Update the unit and functional tests where needed (and add test cases for fee too high, or too low).